requirements.txt
google
python-dotenv
pyinstaller
appdirs
